Fix write and read-write ioctls on certain 64-bit OS
Using int as the operation parameter was wrong. It meant any write or RW
ioctls (which have the MSB set) would be sign-extended to 64-bit when
calling the native API. Depending on the OS, the upper 32 bits were not
cleared prior to comparing with the call number, resulting in
unexpected errors.

Linux and FreeBSD operated properly; I only got the error on Darwin.

(cherry picked from commit feb24153633859502f30bc16b149c78fb1d651b7)

+++ src/network/socket/qnet_unix_p.h
@@ -158,7 +158,7 @@ static inline int qt_safe_connect(int sockfd, const struct sockaddr *addr, QT_SO
 
 // VxWorks' headers specify 'int' instead of '...' for the 3rd ioctl() parameter.
 template <typename T>
-static inline int qt_safe_ioctl(int sockfd, int request, T arg)
+static inline int qt_safe_ioctl(int sockfd, unsigned long request, T arg)
 {
 #ifdef Q_OS_VXWORKS
     return ::ioctl(sockfd, request, (int) arg);
